+++ src/main/java/fr/frogdevelopment/assoplus/dto/SchoolYearDto.java
@@ -4,6 +4,7 @@
 
 package fr.frogdevelopment.assoplus.dto;
 
+import javafx.beans.property.SimpleIntegerProperty;
 import javafx.beans.property.SimpleListProperty;
 import javafx.beans.property.SimpleLongProperty;
 import javafx.beans.property.SimpleStringProperty;
@@ -11,19 +12,19 @@ import javafx.collections.ObservableList;
 
 public class SchoolYearDto implements Dto {
 
-    private final SimpleLongProperty id = new SimpleLongProperty();
+    private final SimpleIntegerProperty id = new SimpleIntegerProperty();
     private final SimpleStringProperty schoolYear = new SimpleStringProperty();
     private final SimpleListProperty<MemberDto> members = new SimpleListProperty<>();
 
-    public Long getId() {
+    public Integer getId() {
         return id.get();
     }
 
-    public SimpleLongProperty idProperty() {
+    public SimpleIntegerProperty idProperty() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id.set(id);
     }
